Hoi loi

+++ Main/Main.py
@@ -9,10 +9,11 @@ from typing import Iterator
 from yt_dlp import YoutubeDL
 from interactions import SlashContext, listen, slash_command, Embed, OptionType, slash_option, ActiveVoiceState, \
     ButtonStyle, ActionRow, Button
+# from interactions.models.discord.channel import ThreadChannel || phát triển code sử dụng module threadchannel
 from interactions.api.events import Startup, VoiceStateUpdate, Component
 from interactions.api.voice.audio import AudioVolume, BaseAudio
 
-Token = os.getenv("Discord_Token_bot_A")
+Token = os.getenv("Discord_Token_Bot_A")
 # Token = os.getenv("Discord_Token_bot_B")
 bot = interactions.Client()
 startup = dt.datetime.utcnow()
@@ -37,25 +38,35 @@ class NaffQueue:
     voice_state: ActiveVoiceState
     _entries: deque
     _item_queued: asyncio.Event
+    loopstate: bool
+    last: BaseAudio
 
     def __init__(self, voice_state: ActiveVoiceState):
         self.voice_state = voice_state
         self._entries = deque()
         self._item_queued = asyncio.Event()
         self.__song_list__ = []
+        self._current_task = None
+        self.loopstate = False
 
     def __len__(self) -> int:
         return len(self._entries)
 
+    def loops(self) -> None:
+        if not self.loopstate:
+            self.loopstate = True
+        elif self.loopstate:
+            self.loopstate = False
+
     def __iter__(self) -> Iterator[BaseAudio]:
         return iter(self._entries)
 
-    def put(self, audio: BaseAudio, ctx: SlashContext) -> None:
-        title = audio.entry['title']
+    def put(self, audio_d: BaseAudio, ctx: SlashContext) -> None:
+        title = audio_d.entry['title']
         titles.append(title)
-        thumbnail = audio.entry['thumbnail']
-        uploader = audio.entry['uploader']
-        duration = audio.entry['duration']
+        thumbnail = audio_d.entry['thumbnail']
+        uploader = audio_d.entry['uploader']
+        duration = audio_d.entry['duration']
         embedmusic_inqueue = Embed(
             title=f" {title}",
             description="ㅤ",
@@ -68,21 +79,21 @@ class NaffQueue:
         embedmusic_inqueue.add_field(name=" Dài:  ", value=f"{duration_hms}", inline=True)
         embedmusic_inqueue.set_thumbnail(url=ctx.author.avatar_url)
         self.__song_list__.insert(0, embedmusic_inqueue)
-        self._entries.append(audio)
+        self._entries.append(audio_d)
         self._item_queued.set()
 
     def get_list(self) -> list:
         return self.__song_list__
 
-    def put_first(self, audio: BaseAudio) -> None:
-        self._entries.appendleft(audio)
+    def put_first(self, audio_d: BaseAudio) -> None:
+        self._entries.appendleft(audio_d)
         self._item_queued.set()
 
     async def pop(self) -> BaseAudio:
         if len(self) == 0:
             await self._item_queued.wait()
         item = self._entries.popleft()
-        self._item_queued.clear()
+        # self._item_queued.clear()
         return item
 
     def pop_no_wait(self) -> BaseAudio:
@@ -104,29 +115,49 @@ class NaffQueue:
         return self._entries[index]
 
     async def __playback_queue(self) -> None:
+        # while self.voice_state.connected:
+        #     if self.voice_state.playing:
+        #         await self.voice_state.wait_for_stopped()
+        #
+        #     if not self.loopstate:
+        #         audio = await self.pop()
+        #         self.last = audio
+        #         print(f">>trang thai lap true {self.loopstate} last {audio.entry['title']} ditucchcfafc")
+        #         embed = self.__song_list__.pop()
+        #         await self.voice_state.channel.send(embed=embed)
+        #         await self.voice_state.channel.send(audio.entry['title'])
+        #     elif self.loopstate:
+        #         self.put(self.last)
+        #         audio = await self.pop()
+        #         print(f">>trang thai lap true {self.loopstate} last {audio.entry['title']}")
+        #
+        #
+        #     await self.voice_state.play(audio)
+        # >> code lặp cần fix << >>> put_at_index tránh việc đang lặp mà thêm bài hát vào
         while self.voice_state.connected:
             if self.voice_state.playing:
                 await self.voice_state.wait_for_stopped()
-
-            audio = await self.pop()
-            if audio is not None:
+            audio_d = await self.pop()
+            if audio_d is not None:
                 embed = self.__song_list__.pop()
                 await self.voice_state.channel.send(embed=embed)
-            await self.voice_state.play(audio)
+            await self.voice_state.play(audio_d)
+
+    async def _stop(self) -> None:
+        await self.voice_state.stop()
 
     async def __call__(self) -> None:
         await self.__playback_queue()
 
     def start(self) -> None:
-        asyncio.create_task(self())
+        if self._current_task is not None:
+            self._current_task.cancel()
+            self._stop()
+        self._current_task = asyncio.create_task(self())
 
     async def skip(self):
-        if len(self._entries) > 0:
-            self._entries.popleft()  # Loại bỏ bài hát đầu tiên khỏi hàng đợi
-
-        if len(self._entries) > 0:
-            next_audio = self._entries[0]  # Lấy bài hát tiếp theo từ hàng đợi
-            await self.voice_state.play(next_audio)  # Bắt đầu phát bài hát tiếp theo
+        next_audio = await self.pop()
+        await self.voice_state.play(next_audio)
 
 
 class YTAudio(AudioVolume):
@@ -203,7 +234,7 @@ def convert_seconds_to_hms(seconds):
 
 
 perm_ck = None
-queues = NaffQueue(None)
+queues = NaffQueue
 titles = []
 
 
@@ -216,18 +247,8 @@ async def _playlist(ctx: SlashContext):
         await ctx.send(f"{x}")
 
 
-@slash_command(name="doi", description="Thêm nhạc vào hàng đợi")
-@interactions.slash_option("song", "Đường dẫns nhạc", 3, True)
-async def _doi(ctx: SlashContext, song: str):
-    print(f"{ctx.guild.name}::{ctx.user.display_name} > DOI |>>{ctx} \n")
-    global queues
-    if not ctx.responded:
-        await ctx.defer()
-    audio = await YTAudio.from_url(song, stream=True)
-    queues.put(audio, ctx)
-    title = audio.entry['title']
-    titles.append(title)
-    await ctx.send(f"Thêm {title} vào hàng đợi thành công", ephemeral=True)
+audio = None
+channelss = None
 
 
 @slash_command(name="play", description="chơi nhạc")
@@ -235,94 +256,103 @@ async def _doi(ctx: SlashContext, song: str):
 async def play(ctx: SlashContext, song: str):
     print(f"{ctx.guild.name}::{ctx.user.display_name} > PLAY |>>{ctx} \n")
     await ctx.defer()
-    global perm_ck, queues, titles
+    global perm_ck, queues, titles, audio, channelss
+    t = True
     titles = []
-    if ctx.voice_state is not None and ctx.voice_state.channel.voice_state.playing is True:
-        audio = await YTAudio.from_url(song, stream=True)
-        title = audio.entry['title']
-        titles.append(title)
-        queues.put(audio, ctx)
-        thumbnail = audio.entry['thumbnail']
-        uploader = audio.entry['uploader']
-        duration = audio.entry['duration']
-        embedmusic_inqueue = Embed(
-            title=f" {title}",
-            description="ㅤ",
-            color=0x5f9afa,
-        )
-        duration_hms = convert_seconds_to_hms(duration)
-        embedmusic_inqueue.set_author('➕ Đã Thêm Vào hàng đợi')
-        embedmusic_inqueue.set_image(thumbnail)
-        embedmusic_inqueue.add_field(name="Upload By:  ", value=f"{uploader}", inline=True)
-        embedmusic_inqueue.add_field(name=" Dài:  ", value=f"{duration_hms}", inline=True)
-        embedmusic_inqueue.set_thumbnail(url=ctx.author.avatar_url)
-        await ctx.send(embed=embedmusic_inqueue)
-    else:
-        perm_ck = ctx.user.id
-        if not ctx.voice_state:
+    if not ctx.voice_state:
+        if ctx.author.voice is not None:
             await ctx.author.voice.channel.connect()
-        audio = await YTAudio.from_url(song, stream=True)
-        title = audio.entry['title']
-        titles.append(title)
-        thumbnail = audio.entry['thumbnail']
-        uploader = audio.entry['uploader']
-        duration = audio.entry['duration']
-        embedmusic = Embed(
-            title=f" {title}",
-            description="ㅤ",
-            color=0x5f9afa,
-        )
-        duration_hms = convert_seconds_to_hms(duration)
-        embedmusic.set_author('📀 Đang Chơi Nhạc')
-        embedmusic.set_image(thumbnail)
-        embedmusic.add_field(name="Upload By:  ", value=f"{uploader}", inline=True)
-        embedmusic.add_field(name=" Dài:  ", value=f"{duration_hms}", inline=True)
-        embedmusic.set_thumbnail(url=ctx.author.avatar_url)
-        hang1 = ActionRow(
-            Button(
-                custom_id="pause_button",
-                style=ButtonStyle.BLUE,
-                label="⏸️ Tạm Dừng",
-            ),
-            Button(
-                custom_id="stop_button",
-                style=ButtonStyle.RED,
-                label="🛑 Dừng ",
-            ),
-            Button(
-                custom_id="resume_button",
-                style=ButtonStyle.GREEN,
-                label="▶️ Tiếp tục",
-            ),
-            Button(
-                style=ButtonStyle.URL,
-                label="Youtube",
-                url=song,
+            channelss = ctx.voice_state.channel.voice_state
+            t = True
+        else:
+            await ctx.send('Bạn phải ở trong 1 kênh thoại', ephemeral=True)
+            t = False
+    if t:
+        if ctx.voice_state is not None and ctx.voice_state.channel.voice_state.playing is True:
+            audio = await YTAudio.from_url(song, stream=True)
+            title = audio.entry['title']
+            titles.append(title)
+            queues.put(audio, ctx)
+            thumbnail = audio.entry['thumbnail']
+            uploader = audio.entry['uploader']
+            duration = audio.entry['duration']
+            embedmusic_inqueue = Embed(
+                title=f" {title}",
+                description="ㅤ",
+                color=0x5f9afa,
             )
-        )
-        hang2 = ActionRow(
-            Button(
-                custom_id="vol_up",
-                style=ButtonStyle.GREEN,
-                label="➕ Tăng Âm Lượng",
-            ),
-            Button(
-                custom_id="vol_down",
-                style=ButtonStyle.GREEN,
-                label="➖ Giảm Âm Lượng",
-            ),
-            Button(
-                custom_id="skip_button",
-                style=ButtonStyle.GREY,
-                label="⏭️ Tiếp theo",
+            duration_hms = convert_seconds_to_hms(duration)
+            embedmusic_inqueue.set_author('➕ Đã Thêm Vào hàng đợi')
+            embedmusic_inqueue.set_image(thumbnail)
+            embedmusic_inqueue.add_field(name="Upload By:  ", value=f"{uploader}", inline=True)
+            embedmusic_inqueue.add_field(name=" Dài:  ", value=f"{duration_hms}", inline=True)
+            embedmusic_inqueue.set_thumbnail(url=ctx.author.avatar_url)
+            await ctx.send(embed=embedmusic_inqueue)
+        else:
+            perm_ck = ctx.user.id
+            audio = await YTAudio.from_url(song, stream=True)
+            title = audio.entry['title']
+            titles.append(title)
+            thumbnail = audio.entry['thumbnail']
+            uploader = audio.entry['uploader']
+            duration = audio.entry['duration']
+            embedmusic = Embed(
+                title=f" {title}",
+                description="ㅤ",
+                color=0x5f9afa,
             )
-        )
-        embedmusic.add_field(name="Tình Đẹp Trai", value="  ")
-        channel = ctx.voice_state.channel.voice_state
-        queues = NaffQueue(channel)
-        queues.put(audio, ctx)
-        queues.start()
-        await ctx.send(embeds=embedmusic, components=[hang1, hang2])
+            duration_hms = convert_seconds_to_hms(duration)
+            embedmusic.set_author('📀 Đang Chơi Nhạc')
+            embedmusic.set_image(thumbnail)
+            embedmusic.add_field(name="Upload By:  ", value=f"{uploader}", inline=True)
+            embedmusic.add_field(name=" Dài:  ", value=f"{duration_hms}", inline=True)
+            embedmusic.set_thumbnail(url=ctx.author.avatar_url)
+            hang1 = ActionRow(
+                Button(
+                    custom_id="pause_button",
+                    style=ButtonStyle.BLUE,
+                    label="⏸️ Tạm Dừng",
+                ),
+                Button(
+                    custom_id="stop_button",
+                    style=ButtonStyle.RED,
+                    label="🛑 Dừng ",
+                ),
+                Button(
+                    custom_id="resume_button",
+                    style=ButtonStyle.GREEN,
+                    label="▶️ Tiếp tục",
+                ),
+                Button(
+                    custom_id="loop_button",
+                    style=ButtonStyle.GREEN,
+                    label="🔂 Lặp lại",
+                )
+            )
+            hang2 = ActionRow(
+                Button(
+                    custom_id="vol_up",
+                    style=ButtonStyle.GREEN,
+                    label="➕ Tăng Âm Lượng",
+                ),
+                Button(
+                    custom_id="vol_down",
+                    style=ButtonStyle.GREEN,
+                    label="➖ Giảm Âm Lượng",
+                ),
+                Button(
+                    custom_id="skip_button",
+                    style=ButtonStyle.GREY,
+                    label="⏭️ Tiếp theo",
+                )
+            )
+            embedmusic.add_field(name="Tình Đẹp Trai", value="  ")
+            queues = None
+            queues = NaffQueue(channelss)
+            queues.put(audio, ctx)
+            queues.start()
+
+            await ctx.send(embeds=embedmusic, components=[hang1, hang2])
 
 
 @listen(Component)
@@ -349,6 +379,14 @@ async def on_component(event: Component):
             case "skip_button":
                 await skip(ctx)
                 print(f"{ctx.guild.name}::{ctx.user.display_name} > {ctx}:skip \n")
+            case "loop_button":
+                await _loop(ctx)
+
+
+async def _loop(ctx: SlashContext):
+    global queues, audio
+    queues.loops()
+    await ctx.send("Đã bật lặp")
 
 
 @slash_command(name="skip", description="Bỏ qua nhạc")
@@ -376,8 +414,9 @@ async def skip(self):
     player = self.bot.get_bot_voice_state(self.guild_id)
     next_item = queues.peek()
     if next_item is not None:
+        print(next_item.entry['title'])
         await player.stop()
-        queues.start()
+        await self.voice_state.wait_for_stopped()
     else:
         await self.send("Hết nhạc trong hàng đợi", ephemeral=True)
 
@@ -401,9 +440,13 @@ async def _stop(ctx):
 
 
 async def _resume(ctx):
+    global queues
     player = ctx.bot.get_bot_voice_state(ctx.guild_id)
     player.resume()
-    await ctx.send('Đã tiếp tục', ephemeral=True)
+    if ctx.voice_state.channel.voice_state.playing is not True:
+        await ctx.send('Đã tiếp tục', ephemeral=True)
+    else:
+        await ctx.send('Không có nhạc đang dừng', ephemeral=True)
 
 
 currvol = 0.5
@@ -412,7 +455,7 @@ currvol = 0.5
 async def _volup(ctx):
     global currvol
     player = ctx.bot.get_bot_voice_state(ctx.guild_id)
-    currvol += 0.2
+    currvol += 0.1
     player.volume = currvol
     await ctx.send('Đã Tăng âm lượng', ephemeral=True)
 
@@ -420,32 +463,44 @@ async def _volup(ctx):
 async def _voldown(ctx):
     global currvol
     player = ctx.bot.get_bot_voice_state(ctx.guild_id)
-    currvol -= 0.2
+    currvol -= 0.1
     player.volume = currvol
     await ctx.send('Đã Giảm âm lượng', ephemeral=True)
 
 
 channels = None
+channel = None
 
 
 @listen(VoiceStateUpdate)
 async def _join(vs: VoiceStateUpdate):
-    global channels
-    channel = None
+    global channels, channel
     if vs.after is not None and vs.after.channel.id == channels:
         channel = await vs.after.guild.create_voice_channel(f"Kênh {vs.after.member.display_name} ")
         await vs.after.member.move(channel.id)
-    if vs.before is not None and vs.before.channel.id == channel.id:
+    if vs.before is not None and (channel is not None and vs.before.channel.id == channel.id):
         await vs.before.guild.delete_channel(channel.id)
+    if vs.before is not None:
+        voice_channel = vs.before.channel
+        members_in_channel = voice_channel.members
+        num_members = len(members_in_channel)
+        if num_members < 3 and vs.bot.get_bot_voice_state(vs.before.guild) is not None:
+            await vs.bot.get_bot_voice_state(vs.before.guild).disconnect()
+
+
+async def get_remaining_members(channeli):
+    members = await channeli.fetch_members()
+    num_members = len(members)
+    return num_members
 
 
 @slash_command(name="setup", description="Đặt kênh voiceS")
 @slash_option(name="channel", description="Chọn kênh", opt_type=OptionType.CHANNEL, required=True)
-async def _setup(ctx: SlashContext, channel: interactions.OptionType.CHANNEL):
+async def _setup(ctx: SlashContext, channeli: interactions.OptionType.CHANNEL):
     global channels
     print(f"{ctx.guild.name}::{ctx.user.display_name} > SETUP |>> {ctx} \n")
-    channels = channel.id
-    await ctx.send(f"đã đặt kênh {channel.name} thành kênh voiceS")
+    channels = channeli.id
+    await ctx.send(f"đã đặt kênh {channeli.name} thành kênh voiceS")
 
 
 bot.start(Token)
